requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
moviepy>=1.0.3
requests>=2.32.0
pydantic>=2.9.0
python-multipart>=0.0.9
google-auth>=2.29.0
google-cloud-storage>=2.17.0


